resolution of bug
bug included a crash when building a tower in the Release version of the game (without debug). Debug showed no errors, but online it became apparent that godot has a problem with letting debug "correct" errors silently that would crash prod.

after troubleshooting, believe the cause was a single input triggering two instances of the same function, and some hanging variants. Removed duplication of function call from click and bug is fixed in Release.